CI-enabled #1712 (#1723)
* Unit tests after refactoring


* add MerkleUniqueToken and MerkleUniqueTokenId


* save changes


* add all necessary hedera-node changes


* resolve queries and resolve compilation failures


* services state update


* Unit tests after refactoring (#274)

* Fix most of unit tests but 7


* all but 1


* PR requested changes


* NFTs: Refactored E2E (#273)

* Adapted E2E Scenarios after major refactoring


* Refactored nfts transfers (#277)


#182 ! merging nft transfers


* fix build (#278)


* add swirlds alpha version


* Replace invertible with new structures


* Unfinished work


* Enable function


* Refactoring and replacement


* Ownership Change on Mint


* Unique Token Ownership Transfer on NFT Transfer


* Minor change


* Merge Conflicts and refactoring


* Refactoring


* Association and disassociation


* Minor change


* PR Comments Resolution


* Unit Tests


* remove fc invertible map & rename properties


* Coverage fix and refactoring


* Refactored Unit Tests


* Refactored Unit Tests


* Nfts: Get Account NFT Infos Query

* Get Nft Account Infos Query


* Merge conflicts


* Query Range Response Code Renaming


* Remove unnecessary stuff


* Nit resolution


* FeeSchedules Revert


* FeeSchedules Revert


* Re-enable Sonar


* Fix unit test


* Fix Sonar bugs


* Stabilize CI (add TokenGetAccountNftInfos to ExpectedCustomThrottles)


* Quans comments resolution


* make naming consistent + fix unit tests

+++ hedera-node/src/test/java/com/hedera/test/mocks/TestContextValidator.java
@@ -111,7 +111,7 @@ public enum TestContextValidator implements OptionValidator {
 	}
 
 	@Override
-	public ResponseCodeEnum nftMaxQueryRangeCheck(int start, int end) {
+	public ResponseCodeEnum nftMaxQueryRangeCheck(long start, long end) {
 		throw new UnsupportedOperationException();
 	}
